Move more functions and make sure all are properly referenced in other functions
